embeddings/requirements.txt
autopep8==2.0.4
certifi==2023.7.22
charset-normalizer==3.3.2
filelock==3.13.1
fsspec==2023.10.0
grpcio==1.59.2
grpcio-tools==1.59.2
huggingface-hub==0.17.3
idna==3.4
numpy==1.26.1
packaging==23.2
protobuf==4.25.0
pycodestyle==2.11.1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
safetensors==0.4.0
setuptools==68.2.2
tokenizers==0.14.1
tqdm==4.66.1
transformers==4.34.1
typing_extensions==4.8.0
urllib3==2.0.7
wheel==0.41.2
